fix: Moved the Exceptions folder to Services. Renamed the ClientServiseTest and EmployeeServiceTest classes.
