+++ BankService/EmployeeService.cs
@@ -1,5 +1,5 @@
-﻿using Models.ModelsValidationExceptions;
-using Models;
+﻿using Models;
+using Services.Exceptions;
 using System;
 using System.Collections.Generic;
 using System.Linq;
