+++ BankService/Exceptions/ClientAlreadyExistException.cs
@@ -4,11 +4,11 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Models.ModelsValidationExceptions
+namespace Services.Exceptions
 {
     public class ClientAlreadyExistException : Exception
     {
         public ClientAlreadyExistException(string message)
         : base(message) { }
-    }       
+    }
 }
