+++ BankService/Exceptions/EmployeePositionValidationException.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Models.ModelsValidationExceptions
+namespace Services.Exceptions
 {
     public class EmployeePositionValidationException : Exception
     {
