+++ BankService/Exceptions/PersonSeriesOfPassportValidationException.cs
@@ -4,11 +4,11 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Models.ModelsValidationExceptions
+namespace Services.Exceptions
 {
     public class PersonSeriesOfPassportValidationException : Exception
     {
         public PersonSeriesOfPassportValidationException(string message)
         : base(message) { }
-    }   
+    }
 }
