+++ Models/Client.cs
@@ -1,5 +1,4 @@
-﻿using Models.ModelsValidationExceptions;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
