+++ ServiceTests/EmployeeExceptionsTest.cs
@@ -1,5 +1,4 @@
-﻿using Models.ModelsValidationExceptions;
-using Models;
+﻿using Models;
 using Services;
 using System;
 using System.Collections.Generic;
@@ -7,10 +6,11 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using Xunit;
+using Services.Exceptions;
 
 namespace ServiceTests
 {
-    public class EmployeeExceptionsTest
+    public class EmployeeServiceTest
     {
         [Fact]
